Create foldfilter tool
Note: right now lines could end with whitespace and it is assumed that that whitespace is maintained, which might not be a reasonable assumption. The alternative is instead of keeping track of the number of lines we chopped a sentence into, we keep track of the actual delimiters where we chopped, don't feed those to the wrapped MT system, and glue those back in when reconstructing the single long sentence.